% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.heatmap.R
\name{draw.heatmap}
\alias{draw.heatmap}
\title{Generate heatmap of enrichment matrix}
\usage{
draw.heatmap(mat, main = "")
}
\arguments{
\item{mat}{query result matrix from function \code{get.heat.mat()}}

\item{main}{title of the heatmap, default is ""}
}
\value{
generate heatmap to graphic device and return result from calling
 \code{pheatmap()}.
}
\description{
This function generate the enrichment heatmap using pheatmap package.
}
\examples{
result=query.egset.list(query.gr=query.gr, query.score=NULL,
  eqtl.set.list=eset.list, gene.set=biocarta)$result.table
mat=res.get.heat.mat(result, test.method = "fisher")
draw.heatmap(mat, main="enrichment between tissue/pathway")
}
\keyword{heatmap}
