% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{eqtlSet}
\alias{eqtlSet}
\title{eqtlSet Class}
\usage{
eqtlSet(object)
}
\description{
eqtlSet Class contains information for eqtl-gene association, gene identifier, position of SNPs, etc.
}
\details{
\describe{
   \item{tissue}{character; name of the cell/tissue of the eQTL study.}

\item{snp.id}{character; name of the SNPs}

\item{snp.gr}{GenomicRanges; position of the SNPs}

\item{gene}{character; gene identifier}
 }
}
