% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.heatmap.R
\name{export.heatmap}
\alias{export.heatmap}
\title{Export heatmap to file}
\usage{
export.heatmap(mat, file, main = "")
}
\arguments{
\item{mat}{query result from function query.egset.list()}

\item{file}{path to the image file}

\item{main}{title of the heatmap, default is ""}
}
\description{
This function exports the enrichment heatmap to a specific file using pheatmap package.
}
\examples{
export.heatmap(mat, file="result.heatmap.png", main="enrichment between tissue/pathway")
}
\keyword{heatmap}
