% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{geneSet}
\alias{geneSet}
\title{geneSet Class}
\usage{
geneSet(object)
}
\description{
geneSet Class contains information for names of gene sets and a list of gene sets
}
\details{
\describe{
   \item{description}{character; additional information for gene sets, such as names, URLs, a short description, etc.}

\item{gene.set}{list; a list of gene sets; each member of the list is a vector containing a group of gene identifiers}

}
}
