% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.geneset.description.R
\name{get.geneset.description}
\alias{get.geneset.description}
\title{Extract gene set description from geneSet object}
\usage{
get.geneset.description(geneset, geneset.ids)
}
\arguments{
\item{geneset}{A \code{geneSet} object}

\item{geneset.ids}{Character; the names of the gene sets.}
}
\value{
a vector of gene set description from \code{geneSet} description 
slot
}
\description{
This function extracts the gene set description from geneSet object.
}
\examples{
query.result <- query.egset.list(query.gr=query.gr, query.score=NULL,
  eqtl.set.list=eset.list, gene.set=biocarta)
head(query.result$result.table)
query.result$cover.gene
}
