% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.heat.mat.R
\name{get.heat.mat}
\alias{get.heat.mat}
\title{Extract tissue/geneset enrichment matrix from query result}
\usage{
get.heat.mat(res, test.method = c("glm", "fisher", "hypergeom"),
  filter.quantile = 0.5, min.ptw.gene = 30)
}
\arguments{
\item{res}{query result from function query.egset.list()}

\item{test.method}{Choose which enrichment test should be used to retrive p-values from. Options include:"glm"(logistic regression),"fisher"(fisher exact test) and "hypergeom"(hypergeometric test)}

\item{filter.quantile}{Filter option; choose the max quantile of all p-values being kept in the matrix; default is 0.5, which means p-values larger than median p-values are discarded}

\item{min.ptw.gene}{Filter option; minimum number of genes in a pathway; default is 30 (pathway with <30 genes are not included in the matrix)}
}
\description{
This function extracts the enrichment matrix from eQTL list query result. The rows of the matrixs are pathways; and the columns of the matrixs are tissues/cell lines of the eQTL sets. P-Values from enrichment tests are summarized in this matrix
}
\examples{
get.heat.mat(res, test.method="fisher")
}
\keyword{result}
