% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heatmap.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Generate heatmap of enrichment matrix}
\usage{
plot.heatmap(mat, main = "", silent = T)
}
\arguments{
\item{mat}{query result matrix from function \code{get.heat.mat()}}

\item{main}{title of the heatmap, default is ""}
}
\description{
This function generate the enrichment heatmap using pheatmap package.
}
\examples{
result=query.egset.list(query.gr=query.gr, query.score=NULL,
  eqtl.set=eqtl.set.list, gene.set=biocarta)
mat=get.heat.mat(result, test.method = "fisher")
plot.heatmap(mat, main="enrichment between tissue/pathway")
}
\keyword{heatmap}
