% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wordcloud.R
\name{plot.wordcloud}
\alias{plot.wordcloud}
\title{Plot word cloud to file}
\usage{
plot.wordcloud(mat, min.freq.tissue = 5, min.freq.gset = 5,
  max.words = 50)
}
\arguments{
\item{min.freq.tissue}{minimum frequency of tissue/cell to be plotted in the word cloud}

\item{min.freq.gset}{minimum frequency of geneset to be plotted in the word cloud}

\item{result}{query result from function \code{query.egset.list()}}
}
\description{
This function plot the enrichment heatmap to a specific file using pheatmap package.
}
\examples{
result=query.egset.list(query.gr=query.gr, query.score=NULL,
  eqtl.set=eqtl.set.list, gene.set=biocarta)
mat=get.heat.mat(result, test.method = "fisher")
plot.wordcloud(mat, min.freq.tissue=5, min.freq.gset=5, max.words=50)
}
\keyword{cloud}
\keyword{word}
