% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.egset.R
\name{query.egset}
\alias{query.egset}
\title{Query enrichment in geneset through one eQTL set.}
\usage{
query.egset(query.gr, query.score, eqtl.set, gene.set, verbose = FALSE)
}
\arguments{
\item{query.gr}{a GenomicRange object, representing query regions}

\item{query.score}{optional, set to NULL if the regions are not ordered.}

\item{eqtl.set}{an eqtlSet object; the eQTL set to be queried against}

\item{gene.set}{an object of geneSet class; the gene set to be tested}

\item{verbose}{bool; whether to show eqtlSet/geneSet summary information;
default is FALSE}
}
\value{
a list; \code{result.table} is the major result table showing
enrichment assessment;
 \code{cover.gene} is the vector showing the genes from the eqtl Sets
 covered by the query region(s)
}
\description{
This function perform enrichment test between one eQTL set and a group of
 gene sets
}
\examples{
#build one eqtlset
skin.eset <- eset.list$Skin
#query one egset
res.one <- query.egset(
  query.gr=query.gr,
  query.score=NULL,
  eqtl.set=skin.eset,
  gene.set=biocarta
)
#enrichment result table
res.one$result.table
#all the genes associated with eQTLs covered by the query region
res.one$cover.gene
}
