% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.one.set.R
\name{query.one.set}
\alias{query.one.set}
\title{Query enrichment in geneset through one eQTL set and one gene set}
\usage{
query.one.set(query.gr, query.score = NULL, eqtl.set, set.j, q.all, n.snp.t,
  n.gene.all)
}
\arguments{
\item{query.gr}{a GenomicRange object, representing query regions}

\item{query.score}{optional, set to NULL if the regions are not ordered.}

\item{eqtl.set}{an eqtlSet object; the eQTL set to be queried against}

\item{set.j}{a character vector; a set of genes}

\item{q.all}{integer}

\item{n.snp.t}{integer}
}
\description{
This function perform enrichment test between one eQTL set and one gene set
}
\examples{
#to be added
}
