% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.get.tissue.degree.R
\name{res.get.tissue.degree}
\alias{res.get.tissue.degree}
\title{Extract tissue degree from query result}
\usage{
res.get.tissue.degree(res, eqtl.set.list)
}
\arguments{
\item{res}{query result from function query.egset.list()}

\item{eqtl.set.list}{a list of eqtlSet; each member should be an eqtlSet
object}
}
\value{
a list; \code{gene.tissue.map} shows mapping:gene<->tissue;
\code{gene.tissue.degree} shows tissue degree for each gene;
\code{mean.tissue.degree} shows the average tissue digree for each pathway
 in the result table
}
\description{
This function extracts the tissue degree from eQTL list query result for
each pathway.
}
\examples{
result=query.egset.list(query.gr=query.gr, query.score=NULL,
  eqtl.set.list=eset.list, gene.set=biocarta)$result.table
tissue.degree=res.get.tissue.degree(result, eset.list)
head(tissue.degree$gene.tissue.map)
head(tissue.degree$gene.tissue.degree)
head(tissue.degree$mean.tissue.degree)
}
\keyword{result}
