% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.order.by.tissue.R
\name{res.order.by.tissue}
\alias{res.order.by.tissue}
\title{Order result by tissue-specificity}
\usage{
res.order.by.tissue(result, tissue.list)
}
\arguments{
\item{result}{query result from function query.egset.list()}

\item{tissue.list}{a list of tissues to be re-ordered}
}
\description{
This function re-rank the query result by the tissue specificity (from query.tissue())
}
\examples{
#to be added
}
