% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.geneid.R
\name{check.geneid}
\alias{check.geneid}
\title{check compatibility of gene identifiers between eQTL set and gene set}
\usage{
check.geneid(e.set, g.set)
}
\arguments{
\item{e.set}{an eqtlSet object; the eQTL set to be queried against}

\item{g.set}{an object of geneSet class; the gene set to be tested}
}
\description{
This function perform enrichment test between one eQTL set and one gene set
}
\examples{
#to be added
}
