% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allClasses.R
\docType{class}
\name{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet}
\alias{tissue,eqtlSet-method}
\alias{tissue,}
\alias{eqtlSet-method}
\alias{eqtlId,eqtlSet-method}
\alias{eqtlId,}
\alias{eqtlSet-method}
\alias{eqtlRange,eqtlSet-method}
\alias{eqtlRange,}
\alias{eqtlSet-method}
\alias{eqtlGene,eqtlSet-method}
\alias{eqtlGene,}
\alias{eqtlSet-method}
\title{Method tissue.}
\usage{
tissue(x)

eqtlId(x)

eqtlRange(x)

eqtlGene(x)

\S4method{tissue}{eqtlSet}(x)

\S4method{eqtlId}{eqtlSet}(x)

\S4method{eqtlRange}{eqtlSet}(x)

\S4method{eqtlGene}{eqtlSet}(x)
}
\description{
Method tissue.

Method eqtlId.

Method eqtlRange.

Method eqtlGene.
}
\section{Slots}{

\describe{
\item{\code{tissue}}{character; name of the cell/tissue of the eQTL study}

\item{\code{snp.id}}{character; name of the SNPs}

\item{\code{snp.gr}}{GenomicRanges; position of the SNPs}

\item{\code{gene}}{character; gene identifier}
}}

\examples{
brain.file = system.file("extdata", "eqtl/brain.gtex.txt", 
    package = "loci2path")
tab = read.table(brain.file, stringsAsFactors = FALSE, header = TRUE)
snp.gr = GRanges(seqnames = Rle(tab$snp.chr), 
    ranges=IRanges(start=tab$snp.pos, 
    width=1))
brain.eset = eqtlSet(tissue="brain",
    snp.id=tab$snp.id,
    snp.gr=snp.gr,
    gene=as.character(tab$gene.entrez.id))
}
