% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{eqtlSet-class}
\alias{eqtlSet-class}
\alias{eqtlSet}
\title{eqtlSet Class}
\description{
eqtlSet Class contains information for eqtl-gene association,
 gene identifier, position of SNPs, etc.
}
\section{Slots}{

\describe{
\item{\code{tissue}}{character; name of the cell/tissue of the eQTL study}

\item{\code{snp.id}}{character; name of the SNPs}

\item{\code{snp.gr}}{GenomicRanges; position of the SNPs}

\item{\code{gene}}{character; gene identifier}
}}

\examples{
brain.file = system.file("extdata", "eqtl/brain.gtex.txt", 
    package = "loci2path")
tab = read.table(brain.file, stringsAsFactors = FALSE, header = TRUE)
snp.gr = GRanges(seqnames = Rle(tab$snp.chr), 
    ranges=IRanges(start=tab$snp.pos, 
    width=1))
brain.eset = eqtlSet(tissue="brain",
    snp.id=tab$snp.id,
    snp.gr=snp.gr,
    gene=as.character(tab$gene.entrez.id))
}
