% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{eqtlSet}
\alias{eqtlSet}
\title{eqtlSet Class}
\usage{
eqtlSet(object)
}
\description{
eqtlSet Class contains information for eqtl-gene association, gene identifier,
position of SNPs, etc.
}
\section{Slots}{

\describe{
\item{\code{tissue}}{character; name of the cell/tissue of the eQTL study}

\item{\code{snp.id}}{character; name of the SNPs}

\item{\code{snp.gr}}{GenomicRanges; position of the SNPs}

\item{\code{gene}}{character; gene identifier}
}}

