% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.heatmap.R
\name{export.heatmap}
\alias{export.heatmap}
\title{Export heatmap to file}
\usage{
export.heatmap(mat, file, main = "", silent = T)
}
\arguments{
\item{mat}{query result matrix from function \code{get.heat.mat()}}

\item{file}{path to the image file}

\item{main}{title of the heatmap, default is ""}

\item{silent}{bool; whether to show heatmap to current display device; default is T}
}
\description{
This function exports the enrichment heatmap to a specific file using pheatmap package.
}
\examples{
result=query.egset.list(query.gr=query.gr, query.score=NULL, eqtl.set=eqtl.set.list, gene.set=biocarta, lr=F)
mat=get.heat.mat(result, test.method = "fisher")
export.heatmap(mat, file="result.heatmap.png", main="enrichment between tissue/pathway", silent=T)
}
\keyword{heatmap}
