% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{geneSet-class}
\alias{geneSet-class}
\alias{geneSet}
\title{geneSet Class}
\description{
geneSet Class contains information for names of gene sets and a list of gene sets
}
\section{Slots}{

\describe{
\item{\code{total.number.gene}}{numeric; the total number of all genes;
This number is used in enrichment tests}

\item{\code{description}}{vector of character; additional information for gene sets,
such as names, URLs, a short description, etc.}

\item{\code{gene.set}}{list;  a list of gene sets; each member of the list is a vector
containing a group of gene identifiers}
}}

