% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.egset.list.R
\name{query.egset.list}
\alias{query.egset.list}
\title{Query enrichment in geneset through multiple eQTL sets.}
\usage{
query.egset.list(query.gr, query.score, eqtl.set.list, gene.set, parallel = F)
}
\arguments{
\item{query.gr}{a GenomicRange object, representing query regions}

\item{query.score}{optional, set to NULL if the regions are not ordered.}

\item{eqtl.set.list}{a list of eqtlSet; each member should be an eqtlSet object}

\item{gene.set}{an object of geneSet class; the gene set to be tested}

\item{parallel}{bool; whether to enable parallel computing;  default is F}
}
\description{
This is the main function for query. The user need to specify
- (1) Query region; if it is ordered, together with the Query score based on which the regions are ranked
- (2) eQTL set list; this is usually more than one eQTL set.
      Only multiple eQTL set derived from different cells/tissues will show cell/tissue specificity
- (3) gene set; the gene sets that enrichment tests would be performed to.
}
\examples{
#to be added
}
