% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.egset.R
\name{query.egset}
\alias{query.egset}
\title{Query enrichment in geneset through one eQTL set.}
\usage{
query.egset(query.gr, query.score, eqtl.set, gene.set, verbose = F)
}
\arguments{
\item{query.gr}{a GenomicRange object, representing query regions}

\item{query.score}{optional, set to NULL if the regions are not ordered.}

\item{eqtl.set}{an eqtlSet object; the eQTL set to be queried against}

\item{gene.set}{an object of geneSet class; the gene set to be tested}

\item{verbose}{bool; whether to show eqtlSet/geneSet summary information; default is F}
}
\description{
This function perform enrichment test between one eQTL set and a group of gene sets
}
\examples{
#to be added
}
